change PagedResult<T>.Total type to int64

+++ src/Shared.Infrastructure/Model/PagedResult.cs
@@ -6,13 +6,13 @@ namespace Shared.Infrastructure.Model
     {
         public List<T> Data { get; set; }
 
-        public int Total { get; set; }
+        public long Total { get; set; }
 
         public int PageIndex { get; set; }
 
         public int PageSize { get; set; }
 
-        public static PagedResult<T> SuccessResult(List<T> data, int pageIndex, int pageSize, int total)
+        public static PagedResult<T> SuccessResult(List<T> data, int pageIndex, int pageSize, long total)
         {
             return new PagedResult<T>
             {
